Separate the backend into two files
Having the "local" and "remote" backends in two files prevents
the "remote" backend from loading Qt, which fixes issues #252 and #258.
